Completed the core functionality of the sys
Merged the front and backend together. Still work to do but all the core features have been achieved. Refinements need to be made such as avoiding the seats being selected twice. Also maybe rather than using a dictionary we could use a Database or CSV to record previous orders as currently when the application is closed all data is overwritten.